+++ extensions/tinyfaces-nft/CHANGELOG.md
@@ -1,3 +1,7 @@
 # TinyFaces NFT Changelog
 
+## [ENSv2 Support] - 2026-09-23
+
+- Resolve owner names through the ENS Universal Resolver
+
 ## [Initial Version] - 2022-09-07